+++ out/src/helpers/gitflowUtils.js
@@ -1,5 +1,6 @@
 'use strict';
-var gitUtils = require('../helpers/gitUtils');
+Object.defineProperty(exports, "__esModule", { value: true });
+const gitUtils = require("../helpers/gitUtils");
 /*
 A branch name can not:
     - Have a path component that begins with "."
@@ -35,7 +36,63 @@ function hasIllegalChars(branchName) {
     else
         return false;
 }
-function startFeature(rootDir, featureName) {
+function checkForBranch(rootDir, branchName) {
+    return gitUtils.getGitPath().then(function (gitExecutable) {
+        return new Promise(function (resolve, reject) {
+            let options = { cwd: rootDir };
+            let spawn = require('child_process').spawn;
+            let ls = spawn(gitExecutable, ['show-ref', '--verify', '--quiet', 'refs/heads/' + branchName], options);
+            let log = '';
+            let error = '';
+            ls.stdout.on('data', function (data) {
+                log += data + '\n';
+            });
+            ls.stderr.on('data', function (data) {
+                error += data;
+            });
+            ls.on('exit', function (code) {
+                if (code > 0) {
+                    reject(error);
+                    return;
+                }
+                var message = log;
+                if (code === 0 && error.length > 0)
+                    message += '\n\n' + error;
+                resolve(message);
+            });
+        });
+    });
+}
+exports.checkForBranch = checkForBranch;
+function initializeWithDefaults(rootDir) {
+    return gitUtils.getGitPath().then(function (gitExecutable) {
+        return new Promise(function (resolve, reject) {
+            let options = { cwd: rootDir };
+            let spawn = require('child_process').spawn;
+            let ls = spawn(gitExecutable, ['checkout', '-b', 'develop'], options);
+            let log = '';
+            let error = '';
+            ls.stdout.on('data', function (data) {
+                log += data + '\n';
+            });
+            ls.stderr.on('data', function (data) {
+                error += data + '\n';
+            });
+            ls.on('exit', function (code) {
+                if (code > 0) {
+                    reject(error);
+                    return;
+                }
+                var message = log;
+                if (code === 0 && error.length > 0)
+                    message += '\n\n' + error;
+                resolve(message);
+            });
+        });
+    });
+}
+exports.initializeWithDefaults = initializeWithDefaults;
+function startFeature(rootDir, featureName, baseBranch) {
     return gitUtils.getGitPath().then(function (gitExecutable) {
         return new Promise(function (resolve, reject) {
             featureName = featureName.replace(/ /g, '_');
@@ -49,25 +106,27 @@ function startFeature(rootDir, featureName) {
                     '\t- End with ".lock"\n' +
                     '\t- Contain a "\" (backslash))');
             else {
-                var options = { cwd: rootDir };
-                var spawn = require('child_process').spawn;
-                var ls = spawn(gitExecutable, ['flow', 'feature', 'start', featureName], options);
-                var log_1 = '';
-                var error_1 = '';
+                if (!baseBranch)
+                    baseBranch = 'develop';
+                let options = { cwd: rootDir };
+                let spawn = require('child_process').spawn;
+                let ls = spawn(gitExecutable, ['checkout', '-b', 'feature/' + featureName, baseBranch], options);
+                let log = '';
+                let error = '';
                 ls.stdout.on('data', function (data) {
-                    log_1 += data + '\n';
+                    log += data + '\n';
                 });
                 ls.stderr.on('data', function (data) {
-                    error_1 += data;
+                    error += data;
                 });
                 ls.on('exit', function (code) {
                     if (code > 0) {
-                        reject(error_1);
+                        reject(error);
                         return;
                     }
-                    var message = log_1;
-                    if (code === 0 && error_1.length > 0)
-                        message += '\n\n' + error_1;
+                    var message = log;
+                    if (code === 0 && error.length > 0)
+                        message += '\n\n' + error;
                     resolve(message);
                 });
             }
@@ -78,32 +137,32 @@ exports.startFeature = startFeature;
 function finishFeature(rootDir) {
     return gitUtils.getGitPath().then(function (gitExecutable) {
         return new Promise(function (resolve, reject) {
-            var options = { cwd: rootDir };
-            var spawn = require('child_process').spawn;
-            var ls1 = spawn(gitExecutable, ['flow', 'feature', 'list'], options);
+            let options = { cwd: rootDir };
+            let spawn = require('child_process').spawn;
+            let ls1 = spawn(gitExecutable, ['rev-parse', '--abbrev-ref', 'HEAD'], options);
             var branchData = '';
-            var features = [];
             var inFeature = false;
             var currentBranch = '';
+            let log = '';
+            let error = '';
             ls1.stdout.on('data', function (data) {
                 branchData += data;
             });
-            ls1.on('exit', function (data) {
-                features = branchData.replace(/ /g, '').trim().split('\n');
-                features.forEach(function (element) {
-                    if (element.indexOf('*') === 0) {
-                        inFeature = true;
-                        currentBranch = element.substring(1);
-                        return;
-                    }
-                });
+            ls1.stderr.on('data', function (data) {
+                error += data;
+            });
+            ls1.on('exit', function (code) {
+                if (code > 0) {
+                    reject(error);
+                    return;
+                }
+                currentBranch = branchData.replace(/ /g, '').trim().split('\n')[0];
+                inFeature = currentBranch.startsWith('feature/');
                 if (!inFeature) {
                     reject('Not currently on a Feature branch');
                     return;
                 }
-                var ls2 = spawn(gitExecutable, ['flow', 'feature', 'finish', currentBranch], options);
-                var log = '';
-                var error = '';
+                let ls2 = spawn(gitExecutable, ['checkout', 'develop'], options);
                 ls2.stdout.on('data', function (data) {
                     log += data + '\n';
                 });
@@ -115,10 +174,32 @@ function finishFeature(rootDir) {
                         reject(error);
                         return;
                     }
-                    var message = log;
-                    if (code === 0 && error.length > 0)
-                        message += '\n\n' + error;
-                    resolve(message);
+                    let ls3 = spawn(gitExecutable, ['merge', '--no-ff', currentBranch], options);
+                    ls3.stdout.on('data', function (data) {
+                        log += data + '\n';
+                    });
+                    ls3.stderr.on('data', function (data) {
+                        error += data;
+                    });
+                    ls3.on('exit', function (code) {
+                        let ls4 = spawn(gitExecutable, ['branch', '-d', currentBranch], options);
+                        ls4.stdout.on('data', function (data) {
+                            log += data + '\n';
+                        });
+                        ls4.stderr.on('data', function (data) {
+                            error += data;
+                        });
+                        ls4.on('exit', function (code) {
+                            if (code > 0) {
+                                reject(error);
+                                return;
+                            }
+                            var message = log;
+                            if (code === 0 && error.length > 0)
+                                message += '\n\n' + error;
+                            resolve(message);
+                        });
+                    });
                 });
             });
         });
@@ -139,25 +220,25 @@ function startRelease(rootDir, releaseName) {
                     '\t- End with ".lock"\n' +
                     '\t- Contain a "\" (backslash))');
             else {
-                var options = { cwd: rootDir };
-                var spawn = require('child_process').spawn;
-                var ls = spawn(gitExecutable, ['flow', 'release', 'start', releaseName], options);
-                var log_2 = '';
-                var error_2 = '';
+                let options = { cwd: rootDir };
+                let spawn = require('child_process').spawn;
+                let ls = spawn(gitExecutable, ['checkout', '-b', 'release/' + releaseName, 'develop'], options);
+                let log = '';
+                let error = '';
                 ls.stdout.on('data', function (data) {
-                    log_2 += data + '\n';
+                    log += data + '\n';
                 });
                 ls.stderr.on('data', function (data) {
-                    error_2 += data;
+                    error += data;
                 });
                 ls.on('exit', function (code) {
                     if (code > 0) {
-                        reject(error_2);
+                        reject(error);
                         return;
                     }
-                    var message = log_2;
-                    if (code === 0 && error_2.length > 0)
-                        message += '\n\n' + error_2;
+                    var message = log;
+                    if (code === 0 && error.length > 0)
+                        message += '\n\n' + error;
                     resolve(message);
                 });
             }
@@ -168,32 +249,32 @@ exports.startRelease = startRelease;
 function finishRelease(rootDir, releaseTag) {
     return gitUtils.getGitPath().then(function (gitExecutable) {
         return new Promise(function (resolve, reject) {
-            var options = { cwd: rootDir };
-            var spawn = require('child_process').spawn;
-            var ls1 = spawn(gitExecutable, ['flow', 'release', 'list'], options);
+            let options = { cwd: rootDir };
+            let spawn = require('child_process').spawn;
+            let ls1 = spawn(gitExecutable, ['rev-parse', '--abbrev-ref', 'HEAD'], options);
             var branchData = '';
-            var releases = [];
             var inRelease = false;
             var currentBranch = '';
+            let log = '';
+            let error = '';
             ls1.stdout.on('data', function (data) {
                 branchData += data;
             });
-            ls1.on('exit', function (data) {
-                releases = branchData.replace(/ /g, '').trim().split('\n');
-                releases.forEach(function (element) {
-                    if (element.indexOf('*') === 0) {
-                        inRelease = true;
-                        currentBranch = element.substring(1);
-                        return;
-                    }
-                });
+            ls1.stderr.on('data', function (data) {
+                error += data;
+            });
+            ls1.on('exit', function (code) {
+                if (code > 0) {
+                    reject(error);
+                    return;
+                }
+                currentBranch = branchData.replace(/ /g, '').trim().split('\n')[0];
+                inRelease = currentBranch.startsWith('release/');
                 if (!inRelease) {
                     reject('Not currently on a Release branch');
                     return;
                 }
-                var ls2 = spawn(gitExecutable, ['flow', 'release', 'finish', currentBranch, '-m ' + releaseTag], options);
-                var log = '';
-                var error = '';
+                let ls2 = spawn(gitExecutable, ['checkout', 'master'], options);
                 ls2.stdout.on('data', function (data) {
                     log += data + '\n';
                 });
@@ -205,10 +286,54 @@ function finishRelease(rootDir, releaseTag) {
                         reject(error);
                         return;
                     }
-                    var message = log;
-                    if (code === 0 && error.length > 0)
-                        message += '\n\n' + error;
-                    resolve(message);
+                    let ls3 = spawn(gitExecutable, ['merge', '--no-ff', currentBranch], options);
+                    ls3.stdout.on('data', function (data) {
+                        log += data + '\n';
+                    });
+                    ls3.stderr.on('data', function (data) {
+                        error += data;
+                    });
+                    ls3.on('exit', function (code) {
+                        let ls4 = spawn(gitExecutable, ['checkout', 'develop'], options);
+                        ls4.stdout.on('data', function (data) {
+                            log += data + '\n';
+                        });
+                        ls4.stderr.on('data', function (data) {
+                            error += data;
+                        });
+                        ls4.on('exit', function (code) {
+                            if (code > 0) {
+                                reject(error);
+                                return;
+                            }
+                            let ls5 = spawn(gitExecutable, ['merge', '--no-ff', currentBranch], options);
+                            ls5.stdout.on('data', function (data) {
+                                log += data + '\n';
+                            });
+                            ls5.stderr.on('data', function (data) {
+                                error += data;
+                            });
+                            ls5.on('exit', function (code) {
+                                let ls6 = spawn(gitExecutable, ['branch', '-d', currentBranch], options);
+                                ls6.stdout.on('data', function (data) {
+                                    log += data + '\n';
+                                });
+                                ls6.stderr.on('data', function (data) {
+                                    error += data;
+                                });
+                                ls6.on('exit', function (code) {
+                                    if (code > 0) {
+                                        reject(error);
+                                        return;
+                                    }
+                                    var message = log;
+                                    if (code === 0 && error.length > 0)
+                                        message += '\n\n' + error;
+                                    resolve(message);
+                                });
+                            });
+                        });
+                    });
                 });
             });
         });
@@ -229,25 +354,25 @@ function startHotfix(rootDir, hotfixName) {
                     '\t- End with ".lock"\n' +
                     '\t- Contain a "\" (backslash))');
             else {
-                var options = { cwd: rootDir };
-                var spawn = require('child_process').spawn;
-                var ls = spawn(gitExecutable, ['flow', 'hotfix', 'start', hotfixName], options);
-                var log_3 = '';
-                var error_3 = '';
+                let options = { cwd: rootDir };
+                let spawn = require('child_process').spawn;
+                let ls = spawn(gitExecutable, ['checkout', '-b', 'hotfix/' + hotfixName, 'master'], options);
+                let log = '';
+                let error = '';
                 ls.stdout.on('data', function (data) {
-                    log_3 += data + '\n';
+                    log += data + '\n';
                 });
                 ls.stderr.on('data', function (data) {
-                    error_3 += data;
+                    error += data;
                 });
                 ls.on('exit', function (code) {
                     if (code > 0) {
-                        reject(error_3);
+                        reject(error);
                         return;
                     }
-                    var message = log_3;
-                    if (code === 0 && error_3.length > 0)
-                        message += '\n\n' + error_3;
+                    var message = log;
+                    if (code === 0 && error.length > 0)
+                        message += '\n\n' + error;
                     resolve(message);
                 });
             }
@@ -258,32 +383,32 @@ exports.startHotfix = startHotfix;
 function finishHotfix(rootDir, hotfixTag) {
     return gitUtils.getGitPath().then(function (gitExecutable) {
         return new Promise(function (resolve, reject) {
-            var options = { cwd: rootDir };
-            var spawn = require('child_process').spawn;
-            var ls1 = spawn(gitExecutable, ['flow', 'hotfix', 'list'], options);
+            let options = { cwd: rootDir };
+            let spawn = require('child_process').spawn;
+            let ls1 = spawn(gitExecutable, ['rev-parse', '--abbrev-ref', 'HEAD'], options);
             var branchData = '';
-            var hotfixes = [];
             var inHotfix = false;
             var currentBranch = '';
+            let log = '';
+            let error = '';
             ls1.stdout.on('data', function (data) {
                 branchData += data;
             });
-            ls1.on('exit', function (data) {
-                hotfixes = branchData.replace(/ /g, '').trim().split('\n');
-                hotfixes.forEach(function (element) {
-                    if (element.indexOf('*') === 0) {
-                        inHotfix = true;
-                        currentBranch = element.substring(1);
-                        return;
-                    }
-                });
+            ls1.stderr.on('data', function (data) {
+                error += data;
+            });
+            ls1.on('exit', function (code) {
+                if (code > 0) {
+                    reject(error);
+                    return;
+                }
+                currentBranch = branchData.replace(/ /g, '').trim().split('\n')[0];
+                inHotfix = currentBranch.startsWith('hotfix/');
                 if (!inHotfix) {
                     reject('Not currently on a Hotfix branch');
                     return;
                 }
-                var ls2 = spawn(gitExecutable, ['flow', 'hotfix', 'finish', currentBranch, '-m ' + hotfixTag], options);
-                var log = '';
-                var error = '';
+                let ls2 = spawn(gitExecutable, ['checkout', 'master'], options);
                 ls2.stdout.on('data', function (data) {
                     log += data + '\n';
                 });
@@ -295,10 +420,54 @@ function finishHotfix(rootDir, hotfixTag) {
                         reject(error);
                         return;
                     }
-                    var message = log;
-                    if (code === 0 && error.length > 0)
-                        message += '\n\n' + error;
-                    resolve(message);
+                    let ls3 = spawn(gitExecutable, ['merge', '--no-ff', currentBranch], options);
+                    ls3.stdout.on('data', function (data) {
+                        log += data + '\n';
+                    });
+                    ls3.stderr.on('data', function (data) {
+                        error += data;
+                    });
+                    ls3.on('exit', function (code) {
+                        let ls4 = spawn(gitExecutable, ['checkout', 'develop'], options);
+                        ls4.stdout.on('data', function (data) {
+                            log += data + '\n';
+                        });
+                        ls4.stderr.on('data', function (data) {
+                            error += data;
+                        });
+                        ls4.on('exit', function (code) {
+                            if (code > 0) {
+                                reject(error);
+                                return;
+                            }
+                            let ls5 = spawn(gitExecutable, ['merge', '--no-ff', currentBranch], options);
+                            ls5.stdout.on('data', function (data) {
+                                log += data + '\n';
+                            });
+                            ls5.stderr.on('data', function (data) {
+                                error += data;
+                            });
+                            ls5.on('exit', function (code) {
+                                let ls6 = spawn(gitExecutable, ['branch', '-d', currentBranch], options);
+                                ls6.stdout.on('data', function (data) {
+                                    log += data + '\n';
+                                });
+                                ls6.stderr.on('data', function (data) {
+                                    error += data;
+                                });
+                                ls6.on('exit', function (code) {
+                                    if (code > 0) {
+                                        reject(error);
+                                        return;
+                                    }
+                                    var message = log;
+                                    if (code === 0 && error.length > 0)
+                                        message += '\n\n' + error;
+                                    resolve(message);
+                                });
+                            });
+                        });
+                    });
                 });
             });
         });