+++ out/src/settings/configSettings.js
@@ -1,0 +1,14 @@
+'use strict';
+Object.defineProperty(exports, "__esModule", { value: true });
+class ConfigSettings {
+    constructor() {
+        this.develop = '';
+        this.master = '';
+        this.releases = '';
+        this.hotfixes = '';
+        this.features = '';
+        this.support = '';
+    }
+}
+exports.ConfigSettings = ConfigSettings;
+//# sourceMappingURL=configSettings.js.map