+++ src/helpers/gitflowUtils.ts
@@ -43,7 +43,65 @@ function hasIllegalChars(branchName) {
     else return false;
 }
 
-export function startFeature(rootDir, featureName) {
+export function checkForBranch(rootDir, branchName) {
+    return gitUtils.getGitPath().then(function (gitExecutable) {
+        return new Promise(function (resolve, reject) {
+            let options = { cwd: rootDir };
+            let spawn = require('child_process').spawn;
+            let ls = spawn(gitExecutable, ['show-ref', '--verify', '--quiet', 'refs/heads/' + branchName], options);
+            let log = '';
+            let error = '';
+            ls.stdout.on('data', function (data) {
+                log += data + '\n';
+            });
+            ls.stderr.on('data', function (data) {
+                error += data;
+            });
+            ls.on('exit', function (code) {
+                if(code > 0) {
+                    reject(error);
+                    return;
+                }
+                var message = log;
+                if(code === 0 && error.length > 0)
+                    message += '\n\n' + error;
+                    
+                resolve(message);
+            });
+        });
+    });
+}
+
+export function initializeWithDefaults(rootDir) {
+    return gitUtils.getGitPath().then(function(gitExecutable) {
+        return new Promise(function(resolve, reject) {
+            let options = { cwd: rootDir };
+            let spawn = require('child_process').spawn;
+            let ls = spawn(gitExecutable, ['checkout', '-b', 'develop'], options);
+            let log = '';
+            let error = '';
+            ls.stdout.on('data', function(data) {
+                log += data + '\n';
+            });
+            ls.stderr.on('data', function(data) {
+                error += data + '\n';
+            });
+            ls.on('exit', function(code) {
+                if(code > 0) {
+                    reject(error);
+                    return;
+                }
+                var message = log;
+                if(code === 0 && error.length > 0)
+                    message += '\n\n' + error;
+
+                resolve(message);
+            });
+        });
+    });
+}
+
+export function startFeature(rootDir, featureName, baseBranch) {
     return gitUtils.getGitPath().then(function (gitExecutable) {
         return new Promise(function (resolve, reject) {
             featureName = featureName.replace(/ /g, '_');
@@ -58,9 +116,12 @@ export function startFeature(rootDir, featureName) {
                     '\t- End with ".lock"\n' +
                     '\t- Contain a "\" (backslash))');
             else {
+                if(!baseBranch)
+                    baseBranch = 'develop';
+
                 let options = { cwd: rootDir };
                 let spawn = require('child_process').spawn;
-                let ls = spawn(gitExecutable, ['flow', 'feature', 'start', featureName], options);
+                let ls = spawn(gitExecutable, ['checkout', '-b', 'feature/' + featureName, baseBranch], options);
                 let log = '';
                 let error = '';
                 ls.stdout.on('data', function (data) {
@@ -92,30 +153,32 @@ export function finishFeature(rootDir) {
         return new Promise(function(resolve, reject) {
             let options = { cwd: rootDir };
             let spawn = require('child_process').spawn;
-            let ls1 = spawn(gitExecutable, ['flow', 'feature', 'list'], options);
+            let ls1 = spawn(gitExecutable, ['rev-parse', '--abbrev-ref', 'HEAD'], options);
             var branchData = '';
-            var features = [];
             var inFeature = false;
             var currentBranch = '';
+            let log = '';
+            let error = '';
             ls1.stdout.on('data', function (data) {
                 branchData += data;
             });
-            ls1.on('exit', function(data) {
-                features = branchData.replace(/ /g,'').trim().split('\n');
-                features.forEach(element => {
-                    if(element.indexOf('*') === 0) {
-                        inFeature = true;
-                        currentBranch = element.substring(1);
-                        return;
-                    }
-                });
+            ls1.stderr.on('data', function (data) {
+                error += data;
+            });
+            ls1.on('exit', function(code) {
+                if(code > 0) {
+                    reject(error);
+                    return;
+                }
+
+                currentBranch = branchData.replace(/ /g,'').trim().split('\n')[0];
+                inFeature = currentBranch.startsWith('feature/');
+
                 if(!inFeature) {
                     reject('Not currently on a Feature branch');
                     return;
                 }
-                let ls2 = spawn(gitExecutable, ['flow', 'feature', 'finish', currentBranch], options);
-                let log = '';
-                let error = '';
+                let ls2 = spawn(gitExecutable, ['checkout', 'develop'], options);
                 ls2.stdout.on('data', function (data) {
                     log += data + '\n';
                 });
@@ -127,11 +190,34 @@ export function finishFeature(rootDir) {
                         reject(error);
                         return;
                     }
-                    var message = log;
-                    if(code === 0 && error.length > 0)
-                        message += '\n\n' + error;
-                        
-                    resolve(message);
+    
+                    let ls3 = spawn(gitExecutable, ['merge', '--no-ff', currentBranch], options);
+                    ls3.stdout.on('data', function (data) {
+                        log += data + '\n';
+                    });
+                    ls3.stderr.on('data', function (data) {
+                        error += data;
+                    });
+                    ls3.on('exit', function (code) {
+                        let ls4 = spawn(gitExecutable, ['branch', '-d', currentBranch], options);
+                        ls4.stdout.on('data', function (data) {
+                            log += data + '\n';
+                        });
+                        ls4.stderr.on('data', function (data) {
+                            error += data;
+                        });
+                        ls4.on('exit', function (code) {
+                            if(code > 0) {
+                                reject(error);
+                                return;
+                            }
+                            var message = log;
+                            if(code === 0 && error.length > 0)
+                                message += '\n\n' + error;
+                                
+                            resolve(message);
+                        });
+                    });
                 });
             });
         });
@@ -155,7 +241,7 @@ export function startRelease(rootDir, releaseName) {
             else {
                 let options = { cwd: rootDir };
                 let spawn = require('child_process').spawn;
-                let ls = spawn(gitExecutable, ['flow', 'release', 'start', releaseName], options);
+                let ls = spawn(gitExecutable, ['checkout', '-b', 'release/' + releaseName, 'develop'], options);
                 let log = '';
                 let error = '';
                 ls.stdout.on('data', function (data) {
@@ -185,31 +271,32 @@ export function finishRelease(rootDir, releaseTag) {
         return new Promise(function(resolve, reject) {
             let options = { cwd: rootDir };
             let spawn = require('child_process').spawn;
-            let ls1 = spawn(gitExecutable, ['flow', 'release', 'list'], options);
+            let ls1 = spawn(gitExecutable, ['rev-parse', '--abbrev-ref', 'HEAD'], options);
             var branchData = '';
-            var releases = [];
             var inRelease = false;
             var currentBranch = '';
+            let log = '';
+            let error = '';
             ls1.stdout.on('data', function (data) {
                 branchData += data;
             });
-            ls1.on('exit', function(data) {
-                releases = branchData.replace(/ /g,'').trim().split('\n');
-                releases.forEach(element => {
-                    if(element.indexOf('*') === 0) {
-                        inRelease = true;
-                        currentBranch = element.substring(1);
-                        return;
-                    }
-                });
+            ls1.stderr.on('data', function (data) {
+                error += data;
+            });
+            ls1.on('exit', function(code) {
+                if(code > 0) {
+                    reject(error);
+                    return;
+                }
+
+                currentBranch = branchData.replace(/ /g,'').trim().split('\n')[0];
+                inRelease = currentBranch.startsWith('release/');
+                
                 if(!inRelease) {
                     reject('Not currently on a Release branch');
                     return;
                 }
-                
-                let ls2 = spawn(gitExecutable, ['flow', 'release', 'finish', currentBranch, '-m ' + releaseTag], options);
-                let log = '';
-                let error = '';
+                let ls2 = spawn(gitExecutable, ['checkout', 'master'], options);
                 ls2.stdout.on('data', function (data) {
                     log += data + '\n';
                 });
@@ -221,13 +308,59 @@ export function finishRelease(rootDir, releaseTag) {
                         reject(error);
                         return;
                     }
-                    var message = log;
-                    if(code === 0 && error.length > 0)
-                        message += '\n\n' + error;
-                        
-                    resolve(message);
+    
+                    let ls3 = spawn(gitExecutable, ['merge', '--no-ff', currentBranch], options);
+                    ls3.stdout.on('data', function (data) {
+                        log += data + '\n';
+                    });
+                    ls3.stderr.on('data', function (data) {
+                        error += data;
+                    });
+                    ls3.on('exit', function (code) {
+                        let ls4 = spawn(gitExecutable, ['checkout', 'develop'], options);
+                        ls4.stdout.on('data', function (data) {
+                            log += data + '\n';
+                        });
+                        ls4.stderr.on('data', function (data) {
+                            error += data;
+                        });
+                        ls4.on('exit', function (code) {
+                            if(code > 0) {
+                                reject(error);
+                                return;
+                            }
+            
+                            let ls5 = spawn(gitExecutable, ['merge', '--no-ff', currentBranch], options);
+                            ls5.stdout.on('data', function (data) {
+                                log += data + '\n';
+                            });
+                            ls5.stderr.on('data', function (data) {
+                                error += data;
+                            });
+                            ls5.on('exit', function (code) {
+                                let ls6 = spawn(gitExecutable, ['branch', '-d', currentBranch], options);
+                                ls6.stdout.on('data', function (data) {
+                                    log += data + '\n';
+                                });
+                                ls6.stderr.on('data', function (data) {
+                                    error += data;
+                                });
+                                ls6.on('exit', function (code) {
+                                    if(code > 0) {
+                                        reject(error);
+                                        return;
+                                    }
+                                    var message = log;
+                                    if(code === 0 && error.length > 0)
+                                        message += '\n\n' + error;
+                                        
+                                    resolve(message);
+                                });
+                            });
+                        });
+                    });
                 });
-            });  
+            });
         });
     });
 }
@@ -249,7 +382,7 @@ export function startHotfix(rootDir, hotfixName) {
             else {
                 let options = { cwd: rootDir };
                 let spawn = require('child_process').spawn;
-                let ls = spawn(gitExecutable, ['flow', 'hotfix', 'start', hotfixName], options);
+                let ls = spawn(gitExecutable, ['checkout', '-b', 'hotfix/' + hotfixName, 'master'], options);
                 let log = '';
                 let error = '';
                 ls.stdout.on('data', function (data) {
@@ -279,30 +412,32 @@ export function finishHotfix(rootDir, hotfixTag) {
         return new Promise(function(resolve, reject) {
             let options = { cwd: rootDir };
             let spawn = require('child_process').spawn;
-            let ls1 = spawn(gitExecutable, ['flow', 'hotfix', 'list'], options);
+            let ls1 = spawn(gitExecutable, ['rev-parse', '--abbrev-ref', 'HEAD'], options);
             var branchData = '';
-            var hotfixes = [];
             var inHotfix = false;
             var currentBranch = '';
+            let log = '';
+            let error = '';
             ls1.stdout.on('data', function (data) {
                 branchData += data;
             });
-            ls1.on('exit', function(data) {
-                hotfixes = branchData.replace(/ /g,'').trim().split('\n');
-                hotfixes.forEach(element => {
-                    if(element.indexOf('*') === 0) {
-                        inHotfix = true;
-                        currentBranch = element.substring(1);
-                        return;
-                    }
-                });
+            ls1.stderr.on('data', function (data) {
+                error += data;
+            });
+            ls1.on('exit', function(code) {
+                if(code > 0) {
+                    reject(error);
+                    return;
+                }
+
+                currentBranch = branchData.replace(/ /g,'').trim().split('\n')[0];
+                inHotfix = currentBranch.startsWith('hotfix/');
+                
                 if(!inHotfix) {
                     reject('Not currently on a Hotfix branch');
                     return;
                 }
-                let ls2 = spawn(gitExecutable, ['flow', 'hotfix', 'finish', currentBranch, '-m ' + hotfixTag], options);
-                let log = '';
-                let error = '';
+                let ls2 = spawn(gitExecutable, ['checkout', 'master'], options);
                 ls2.stdout.on('data', function (data) {
                     log += data + '\n';
                 });
@@ -314,13 +449,59 @@ export function finishHotfix(rootDir, hotfixTag) {
                         reject(error);
                         return;
                     }
-                    var message = log;
-                    if(code === 0 && error.length > 0)
-                        message += '\n\n' + error;
-                        
-                    resolve(message);
+    
+                    let ls3 = spawn(gitExecutable, ['merge', '--no-ff', currentBranch], options);
+                    ls3.stdout.on('data', function (data) {
+                        log += data + '\n';
+                    });
+                    ls3.stderr.on('data', function (data) {
+                        error += data;
+                    });
+                    ls3.on('exit', function (code) {
+                        let ls4 = spawn(gitExecutable, ['checkout', 'develop'], options);
+                        ls4.stdout.on('data', function (data) {
+                            log += data + '\n';
+                        });
+                        ls4.stderr.on('data', function (data) {
+                            error += data;
+                        });
+                        ls4.on('exit', function (code) {
+                            if(code > 0) {
+                                reject(error);
+                                return;
+                            }
+            
+                            let ls5 = spawn(gitExecutable, ['merge', '--no-ff', currentBranch], options);
+                            ls5.stdout.on('data', function (data) {
+                                log += data + '\n';
+                            });
+                            ls5.stderr.on('data', function (data) {
+                                error += data;
+                            });
+                            ls5.on('exit', function (code) {
+                                let ls6 = spawn(gitExecutable, ['branch', '-d', currentBranch], options);
+                                ls6.stdout.on('data', function (data) {
+                                    log += data + '\n';
+                                });
+                                ls6.stderr.on('data', function (data) {
+                                    error += data;
+                                });
+                                ls6.on('exit', function (code) {
+                                    if(code > 0) {
+                                        reject(error);
+                                        return;
+                                    }
+                                    var message = log;
+                                    if(code === 0 && error.length > 0)
+                                        message += '\n\n' + error;
+                                        
+                                    resolve(message);
+                                });
+                            });
+                        });
+                    });
                 });
-            });  
+            });
         });
     });
 }