+++ src/settings/configSettings.ts
@@ -1,0 +1,10 @@
+'use strict';
+
+export class ConfigSettings {
+    develop: string = '';
+    master: string = '';
+    releases: string = '';
+    hotfixes: string = '';
+    features: string = '';
+    support: string = '';
+}