+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.1] - 2017-8-30
+### Added
+- Ability to initialize git flow repository settings
+
+### Changed
+- Removed need for gitflow or gitflow-avh to be installed prior to using the extension. All commands now run organic git commands
+
+### Fixed
+- Error that occurred when trying to use the extension that claimed there wasn't a git repo in the current or any parent folder
+
 ## [0.0.5] - 2017-1-26
 ### Added
 - This CHANGELOG file.
@@ -33,3 +43,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - First version of extenstion
 - Allowed for creation of features, releases, hotfixes
 - Functionality to convert spaces to _ in branch names
+